API Ensure everything gets flushed when flushing from sake
