+++ src/Core/Flushable.php
@@ -2,8 +2,6 @@
 
 namespace SilverStripe\Core;
 
-use SilverStripe\Control\Middleware\FlushMiddleware;
-
 /**
  * Provides an interface for classes to implement their own flushing functionality
  * whenever flush=1 is requested.
@@ -12,11 +10,9 @@ interface Flushable
 {
 
     /**
-     * This function is triggered early in the request if the "flush" query
-     * parameter has been set. Each class that implements Flushable implements
+     * This function is triggered early in the request if the kernel gets flushed.
+     * Each class that implements Flushable implements
      * this function which looks after it's own specific flushing functionality.
-     *
-     * @see FlushMiddleware
      */
     public static function flush();
 }
